Move test_pjos_disabled to common
cause it just tests WantsOutputs
